backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
sqlalchemy[asyncio]>=2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
redis==5.2.1
python-dotenv==1.0.1
websockets==14.1
anthropic>=0.94.0
httpx==0.28.1
feedparser==6.0.11
apscheduler==3.10.4
loguru==0.7.3
pydantic-settings==2.7.1
lightgbm==4.5.0
scikit-learn==1.5.2
joblib==1.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
webauthn==2.7.1
psutil==6.1.1
